- Modified Allele.cpp to use only the operator< for all comparisons - Added ordering by ID to Group.h - Modified GroupCollection to load groups matching either group names or IDs - Added Information class to get basic information about the database (genomic build, etc) - Added template specialization for std::less for both Region* and Locus* - Prohibited copying of Locus objects - Modified Makefile to get new additions - Added a const_group_iterator class to the Region class - Added ability to iterate over groups of Regions at a particular location -Changed the converter to use Locus* as opposed to Locus objects
git-svn-id: https://svn.ritchielab.psu.edu/svn/projects/biobin@1553 9e62ecf6-770d-0410-a9d7-e62a33e3d092

+++ trunk/src/knowledge/Locus.cpp
@@ -60,7 +60,7 @@ uint Locus::distance(const Locus& other) const{
 }
 
 short Locus::encodeGenotype(uint a1, uint a2) const{
-	if (a1 == -1 || a2 == -1){
+	if (a1 == (uint)-1 || a2 == (uint)-1){
 		return -1;
 	}
 	return a1 * _alleles.size() + a2;
@@ -77,7 +77,7 @@ pair<uint, uint> Locus::decodeGenotype(short encoded_type) const{
 
 
 const string& Locus::getChromStr(short chrom){
-	if (chrom < 0 || chrom > _chrom_list.size()-1){
+	if (chrom < 0 || (uint) chrom > _chrom_list.size()-1){
 		return invalid_chrom;
 	}else{
 		return _chrom_list[chrom];
